reworked audio output to 32-bit float

+++ rendering/audio.h
@@ -55,7 +55,7 @@ extern AudioSenderThread* audio_thread;
 extern QMutex audio_write_lock;
 
 #define audio_ibuffer_size 192000
-extern qint8 audio_ibuffer[audio_ibuffer_size];
+extern float audio_ibuffer[audio_ibuffer_size];
 extern qint64 audio_ibuffer_read;
 extern long audio_ibuffer_frame;
 extern double audio_ibuffer_timecode;
